feat: extract person and player entity from playerlist #165
current implementation can handle players only without history. so when a player changes the team the history will move with the player.
Now the history stays in the team while left_on and joined_on is introduced.
As well as entity player and person.
Fix tests according to new entities.